## Why

`Build/rector/rector.php` assembled its own rule baseline from scratch.
That is what made the Rector 2.6.0 release on 2026-08-03 an org-wide
outage rather than a one-line fix: `SetList::STRICT_BOOLEANS` was
removed upstream and every repo carrying its own set list had to be
repaired individually. The org policy is that
[netresearch/typo3-ci-workflows](https://github.com/netresearch/typo3-ci-workflows)
owns the Rector rule baseline, so the next upstream change is absorbed
once, centrally. This adopts
[v1.3.4](https://github.com/netresearch/typo3-ci-workflows/releases/tag/v1.3.4)
of that shared config, following
[netresearch/t3x-nr-image-optimize#150](https://github.com/netresearch/t3x-nr-image-optimize/pull/150).

## What the shared config now provides

The PHP sets and `phpVersion`, the code-quality sets (`CODE_QUALITY`,
`CODING_STYLE`, `DEAD_CODE`, `EARLY_RETURN`, `INSTANCEOF`,
`PRIVATIZATION`, `TYPE_DECLARATION`, `UP_TO_PHP_82`), the common rule
skips, `importNames()`/`removeUnusedImports()`, and the Rector-specific
PHPStan neon that the package ships for exactly this purpose. All of
that is deleted from the local config.

## What stays local

The `Tests/` path override, because the shared `$projectRoot` default
does not include `Tests/` and `paths()` replaces rather than merges;
`PHPUnitSetList::PHPUNIT_110`; `Typo3SetList::CODE_QUALITY` and
`Typo3SetList::GENERAL`; `Typo3LevelSetList::UP_TO_TYPO3_13` (the v14
set pulls in ComponentFactory/cache migrations this extension does not
need yet); `AddVoidReturnTypeWhereNoReturnRector`, which is not part of
the shared `TYPE_DECLARATION` set; and the three reasoned skips —
`ReadOnlyPropertyRector` on the OpenAI integration test,
`AddErrorCodeToExceptionRector` on the three exceptions whose
constructor takes `($context, ?Throwable $previous)` instead of the
standard `\Throwable` signature, and `Tests/Fuzzy/` where Eris'
namespace functions collide with auto-imports.

## One deliberate removal: ExtEmConfRector

The `ExtEmConfRector` configuration is gone, because under the shared
base it could not fire: the shared config skips `ext_emconf.php`
whenever `$projectRoot` is passed. That is verified rather than assumed
— perturbing the `php` constraint in `ext_emconf.php` and re-running
Rector shows the rule correcting the line under a bare config and
staying silent under this one. Treating `ext_emconf.php` as off-limits
to tooling is consistent across the org (php-cs-fixer `notPath`, PHPStan
`excludePaths`, and this repo's own `.php-cs-fixer.dist.php`), and no
converted repo in the fleet keeps the rule. The constraints the rule
enforced — `typo3: 13.4.0-14.99.99`, `php: 8.2.0-8.99.99` — are already
exactly what `ext_emconf.php` declares, and that file carries a
hand-written comment explaining the intentional TER range which is
better left untouched.

## Source changes

The source diff comes from the sets the shared baseline adds and the
previous config did not have. These are Rector's own output, applied to
a fixpoint, not hand edits.

## Verification

Verified against the toolchain the Rector job actually runs.
`Build/rector/rector.php` is judged on PHP 8.2 in CI
(`rector-php-version: '8.2'` in `.github/workflows/ci.yml`), so this
used `composer config platform.php 8.2.33` with a cold `composer
install` — which resolves phpunit 11.5.56, meaning only the PHPUnit
rules CI can see were applied. The pin was removed again before
committing and `composer.lock` is gitignored.

Checks run: Rector to a fixpoint (`--dry-run` exit 0), `composer
ci:test:php:cgl` clean with Rector still at its fixpoint afterwards,
PHPStan clean at `--memory-limit=1G`, and the unit suite green.

One deviation from `AGENTS.md` worth flagging: it prescribes
`Build/Scripts/runTests.sh`, which needs Docker, and Docker is not
available in the environment this was prepared in. The verification
therefore called `.Build/bin/*` directly under the platform pin, the
same way the reference PRs in this campaign were prepared. CI remains
the authoritative gate.

+++ Tests/Unit/Service/Tool/LogExceptionReaderTest.php
@@ -91,7 +91,7 @@ final class LogExceptionReaderTest extends TestCase
     #[Test]
     public function timeWindowFilters(): void
     {
-        $probe = (int)strtotime('Mon, 06 Jul 2026 01:13:20 +0200');
+        $probe = strtotime('Mon, 06 Jul 2026 01:13:20 +0200');
 
         $entries = $this->reader->read(10, null, $probe - 30, $probe + 30);
